grate: factor out 3d related functions to grate-3d.c
Logical split, no functional change.
